Removes felinid interaction with cocoa (#48066)
Why It's Good For The Game

Redditors are bad for the game, if we can cause some of them to leave, the game will be improved immensely.

Further more this reaction was added during the felinid freeze and should not have been merged by maintainers in the first place.

I was willing to let it slide, but it has now become a problem for me.
Changelog

cl oranges
del: Felinids no longer react to cocoa
/cl