Improves cinematic code (#48193)
* Improves cinematic code

Makes the /obj/screen/cinematic that the cinematics use GC's properly.
Adds a global signal for when a cinematic tries to play. Kills the global list of cinematics.
Makes cinematics no longer loop through the global mob list to make sure that each mob is in fact actually in the global mob list.

* Suggested fixes