Creation of AIs and silicon law changes are now reported in deadchat (#48216)
About The Pull Request

This PR adds deadchat notifications for law changes of AIs and unsynced borgs (listing the full new lawset), uses of AI/Cyborg Upload and the creation of new AIs.
Why It's Good For The Game

The AI is one of the most important players in the round, so it only makes sense that players watching in deadchat are kept up to date on its laws, especially hidden laws since they are never stated.
Changelog

cl
add: Silicon law changes are now reported in deadchat.
add: The creation of new AIs is now reported in deadchat.
/cl