[READY] Roundstart spawning without job landmark puts you on shuttle (#48821)
About The Pull Request

Discovered via paramedics missing their start landmarks, jobs without landmarks would spawn on the last landmark to be loaded in the map. This appears to not have been intended, and they now spawn on the arrivals shuttle if available. If that fails they spawn in the error room.

I am touching job distribution code so I'd like this to be testmerged for a bit before getting merged.
Why It's Good For The Game

This is a bug in an error condition, so it will only show up if there are map bugs, which is why we haven't uncovered it for so long.
Changelog

🆑 JJRcop
fix: Jobs missing spawns will spawn on the arrivals shuttle.
/🆑