Updates all Silicon Law related material costs (#48801)
About The Pull Request

The AI/Cyborg uploads, and AI modules, now require bluespace crystals to construct.
Why It's Good For The Game

Before you say anything I want you to consider the following:

You are an antagonist and you want to subvert the AI. Do you,
A) Break-in to the AI Upload, an extremely risky, obvious method but you have everything you need to subvert the AI without any extra steps?
B) Break-in to secure tech storage, a place with signficiantly less security and not as often checked on, but have to find another way to acquire the modules you need?
C) Wait a good 15-25 minutes for mining and science to do the bare minimum of their job, enter in the back way and print off everything you need with very little risk or leaving behind very little evidence?

If you answered anything but C, I like you and you can fuck my sister but let's be realistic here.
It's not going to """fix""" the issue and I won't pretend it will, but with the rarity of bluespace crystals it does mean that they're not going to be as prevalent or as available as early and may force you to take the riskier strategies if the resources you need don't appear quickly enough and encourage conflict.
Changelog

🆑
balance: AI and Cyborg Upload consoles require bluespace crystals and diamond to print now
balance: Law modules now require bluespace crystals to print now
/🆑