Refactors the crafting component (#48765)
About The Pull Request

Allows the crafting component to not be limited to just mobs but also able to be put on just about any atom, in addition it also gives a bit more settings to play around with like increasing the range of the ingredients search or choosing what to do with an item before it's placed anywhere.
Why It's Good For The Game

Enables greater functionality and may be a precursor PR before implementing automatic crafting machinery as this PR would allow machines to be able to do such a thing.
Changelog

🆑 ma44
refactor: Crafting has now been refactored and allows non mobs to have the ability to craft, at least that's what would come out of it if someone further developed the idea.
/🆑