Random mineral spawns change less turfs. (#49205)
As-is, random mineral spawning changes the turf to the subtype with that mineral, which results in a lot of turf changes I guess. Now it just changes the mineralType of the turf. Unless it's gibtonite or a cave spawner, because the alternatives are worse and those aren't all that common anyway.

🆑
tweak: Random mineral turfs no longer cause large amounts of turf changes on initialization.
/🆑