Secway and janicart key suicides (#49681)
* Update vehicle_key.dm

Adds a unique suicide to the secway key where the user spins, and gibs after a short delay.

* Improvements

Fixes a runtime, adds a little silliness.

* Adds shame result

Grammar changes and "rewards" anyone that successfully fails to commit suicide.

* Adds janicart key suicide

Adds a skillcheck to the suicide that will have varying effects based on the level the user has.

* Adds the return at the end of the file back

Whoops, didn't notice I deleted that I guess. Dunno if that matters, but I have no reason to remove it, so I'm putting it back.