The Infiltrator Sneaksuit Bundle: Armor for (relatively) Cheap (#49608)
About The Pull Request

Redoing the pull request because it's changed so fucking much. Original over here #49550.

The Infiltrator Bundle is a sneaksuit set that allows the wearer the benefits of a decent armor set at a relatively low price of 6TC. The armor vest paired with the sneaksuit will equal a syndicate hardsuit, but only on the chest. It also carries none of the benefits of the hardsuit, such as EVA.

The bundle comes complete with the following items:

Infiltrator Helmet: It's a insidious helmet. Has flashproofing.
Infiltrator Vest: It's roughly equal to an armor vest, but a lot wearker against lasers and disablers.
Blood-red Sneaksuit: It's the highest armored undersuit in the game, at 10 across the board.
Infiltrator Balaclava: Equal to an explorer mask in armor, and has a diagnostic hud built in. It also distorts your voice, so you always come up as 'Unknown' when speaking, even when you have another ID on. Obviously has anti-syngery with the Agent Card, but you're not wearing this to disguise as other people.
Infiltrator Gloves: Special combat gloves that allow you to fireman carry people more quickly.
Sneakboots: They muffle your footsteps completely!
Carrying Case: Storage container for the full suit and guns if you have them. Guns not included.

Also includes the Sleepytime Pajama Bundle for nukies.

Zawo and I have been passing this between us for a while trying to get this into a good state. Hopefully now it's worth an inclusion.
Why It's Good For The Game

A cheaper alternative to the Syndicate Balloon to indicate to everyone around you that you are probably valid.

A recognizable outfit that proves to be useful in minor ways and provide cheap armor for traitors. Nothing this grants is especially powerful on it's own, but it is a useful for traitors wanting to gain an edge at low cost.
Changelog

🆑 NecromancerAnne and zawo
add: The Infiltrator Bundle, an armor kit for 6TC. Murder people in style!
add: Some pajamas for nukies to get plenty of bed rest.
/🆑