Use /tg/'s Discord UserID wiki page over Discord's because Discord's guide sucks. (#50172)
https://tgstation13.org/wiki/How_to_find_your_Discord_User_ID

Somebody who is an experienced discord user will likely already know how to find this, and somebody who is not may not be conformable enabling dev mode to get it. The bot command method is good for people who are on their phone, the url method is good for people on web discord, and desktop app users are likely to already have dev mode on.