[READY] [FREDDY] [BETTY SPAGHETTI] Adds suicide to toy singularity (#50069)
* Toy singularity suicide

I regret nothing.

* tweaks to grammar

* Changes playsound to be at user, not at src

The suicider is eating the toy, so it comes from the suicider, not the toy

* Changes they and have to pronoun helpers

But he has no mouth!
But she has no mouth!
But it has no mouth!
But they have no mouth!

* I forgot ()

* more updates

fixes missing span
singulo now cavity implants the singulo into you
adds new shame check for if you already have a cavity in you while eating singulo
singulos now provide small amounts of nutrition

* it takes 3 seconds for you to choke to death

* Changes sleep to timer, adds another failstate

A huge workaround just to avoid the demonized sleep() proc. Ah well.
Also if you throw the singulo away you choke to death

There might be issues with the regurgitating part of the suicide and the health scanner hud, but that's out of scope for this PR.

* oops

documentation fail

* this should fix documentation

* update documentation on this internal proc

* Update code/game/objects/items/toys.dm

Co-authored-by: moo <11748095+ExcessiveUseOfCobblestone@users.noreply.github.com>