Fix malf AI camera upgrade module (#50249)
Changes a see_invisible value that was being adjusted to be a 
lighting_alpha value that is adjusted when a malf AI uses the camera 
upgrades so that it no longer effectively blinds them.

The lighting_alpha value is the same value that nightvision goggles 
use, so players should be familiar with how things look with it, so 
that they will still be able to discern lighting levels in an area.