Propose requiring use of the policy config system for some features (#50185)
* Propose requiring use of the policy config system for some features

This commit proposes adding the requirement that a developer working on
a ghost role, antag, minor antag or event mob role of any kind must
support the policy configuration system. It also lays out a simple guide
to that system, as well as expectations.

The requirement will go into affect upon merge, any non merged PR at
that time will be required to implement the guidelines

* Update .github/POLICYCONFIG.md

Co-Authored-By: Rohesie <rohesie@gmail.com>

* Update .github/POLICYCONFIG.md

Co-Authored-By: moo <11748095+ExcessiveUseOfCobblestone@users.noreply.github.com>

Co-authored-by: Rohesie <rohesie@gmail.com>
Co-authored-by: moo <11748095+ExcessiveUseOfCobblestone@users.noreply.github.com>