Fix exploit when fireman carrying buckled people (#50364)
Makes you fail when attempting to fireman's carry someone when they are 
buckled to an object, so that you cannot get the off-hand placeholder 
item to enable unintended interactions.

No more long-range body throwing, people teleporting through bluespace 
to the individual with the off-hand placeholder to be slammed onto a 
table, or tabling without needing to regrab the individual.