Fix carp rifts not actually being infinite (#50196)
Fixes carp rifts not spawning in infinite carps upon Space Dragon 
completing his objective. Also renders the carp rift above things like 
doors, as per request. It not already doing this was an oversight, and 
gives Space Dragon winning some weight. Also makes it easier for ghosts 
to use the rift if it shares a tile with something else.