Adds core R&D consoles to the RD's office (#50465)
About The Pull Request

Okay, fine, I agree that heads should be in charge of researching stuff, but the RD having to hoof it to the bridge every time is asinine. "Durr what if the crew breaks in" Let 'em, it's an IC issue.
Why It's Good For The Game

Better usability and less people freaking out about research being impossible now.
Changelog

cl
add: Added an R&D console to the Research Director's office on Meta, Pubby, Box, Delta, and Donut.
/cl