Make it clearler what to do when auto-staled (#50425)
Improves the description of the stale auto issue comment.
About The Pull Request

It is not clear to people how the auto stale process works, and they are having their PR shut because of it.

Some things to note if your PR is autoclosed

A)Preferably, ask a maintainer to reopen the PR, at that point they will decide if your PR was going to be merged and choose to reopen your PR and remove the stale label.

B)Open a new PR with the same branch (less preferred, but fine if you can't get any maintainers attention)

If your PR is marked stale but you have no compile issues or outstanding review, you need to make sure you follow up with maintainers in discord, as it's likely there are outstanding design issues preventing your PR being merged, or no appetite to include the change.