Makes the flash bang ear ring less close to actual tinnitus (#50426)
I could make this effect less shitty if we had a real audio engine but this will have to do.

Adjusts the flashbang ear ring pretty heavily to still have roughly the same effect without sounding that close to actual hearing damage. I have tinnitus myself and the previous sound was uncomfortably close to what real bad bouts of it sound like, and according to some others that have it as well it was too close to theirs and very irritating as a result as well.
Why It's Good For The Game

Less discomfort for everyone.
Changelog

🆑
soundadd: Tweaked the flashbang ear ring which is used in a few places to be less close to actual tinnitus
/🆑