Removes cryo syringes and cryo shotgun darts (#50471)
Purges cryo syringes and cryo shotgun darts, and all references to them from the code.
Why It's Good For The Game

Cryo syringes are, to put it mildly, a blight on chemistry and its balance. Every single design decision around the difficulty of using pyrotechnic chemicals can be sidestepped by heating the ingredients to max temperature and putting them in a single cryo syringe. This leads to instant kill weapons like RDX+teslium or TATP syringes, which mix their ingredients and instantly explode as soon as the syringe makes contact with something.

Nuke ops have to cough up 8 telecrystals for a rocket launcher that causes a roughly similar explosion, plus 4 more for each subsequent rocket, all with the cost of the launcher being a bulky item. Based on this, it seems like it should be inconceivable for chemists to have access to the same firepower with only a syringe gun that fits in a backpack, their chem dispenser available from roundstart and a single techweb node.

Removing these syringes will prevent chemists from evading the dangers of working with explosive chemicals, and should have little to no impact on other applications since very few other reactions exist which need to take place on being injected into a person.

Cryo shotgun darts present an identical problem, and are only not seen in the wild due to their smaller capacity and position far into the techweb making them comparatively undesirable at a point when rapid syringe guns are likely already researched.
Changelog

🆑 Thunder12345
del: Cryo syringes and cryo shotgun darts have been removed from the game
/🆑