Remove exploity way of using pumps and scrubbers (#50542)
I added to portable pumps and scrubbers the same limit t1 canisters have in terms of heat and pressure (5000 K and 50000 Kpa). In hindsight i should've added this in the same PR as the canisters but i forgot about those and i see people use those instead of making t2 and t3 canisters.
Edit: i changed how the damage is calculated, it now uses a clamp() between 5 and 50 so that there will be no more instances where the damage is so low that is discarded and no damage is done
Why It's Good For The Game

Add the intended behaviour for all gas containers thinked in the Tiered Canisters PR to Portable Pumps and Scrubbers
Changelog

cl
add: Added heat and pressure limit to portable pumps and scrubbers
tweak: changed the damage calculation from a max() to a clamp() between 5 and 50
tweak: reduced max pressure output of t3 canisters to the same as t2
/cl