Refocus the map after clicking a goonchat link (#51104)
Clicking a goonchat link will now instantly return focus to the game 
map. This means that the game will no longer eat your next input after 
you clicked a link.

This makes the game feel more responsive and with no downside. Also 
removes an ugly function that was supposed to implement this feature. 
This function was also causing some other issues, for example, it could 
cause you to take a screenshot with f2 although you have rebound it.